Initial Stem Splitting
Implements a basic version of stem splitting by running demucs locally w/ docker. Stem splitting happens upon uploading a transmission.

This needs to be revised still. Stem splitting is a long running operation that the API is not currently spec'd to handle. When stems are being split, it makes the initial POST request block until finished. This is not great in the long run and can be made MUCH better with some smarter design.