Audio Splitting uses BullMQ
Previously, the Transmission POST endpoint would handed audio splitting via Demucs DIRECTLY in the request handler - which caused the response to stall.

To resolve this and provide a more future-proof structure, this PR revises the API spec to make Transmissions more friendly to a long-running audio splitting that happens as a separate job.

API-wise, this means the introduction of a new `splitOperation` field in Transmissions.

Architecturally, this means the introduction of Redis and BullMQ in order to handle audio splitting as a queue of jobs with no concurrency, so that only one song is split at a time. This is nice because it balances the resources required to split over a longer period of time, while still allowing Transmission creation to work quickly.

This structure has much room to expand in the future. When this is eventually not monolithic, I imagine it wouldn't be out of the question to move the actual job execution of audio splitting over to some other elastic container!